+++ Chattokouken Ganbarunoda/AppDataManager.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Globalization;
 using System.IO;
 using System.Text.Json;
 
@@ -7,12 +8,68 @@ namespace CKG
 {
     public readonly record struct SProfileInfo(string Name, int Number);
 
-    public static class ProfileManager
+    public static class AppDataManager
     {
+        private const string CONFIG_FILE_NAME = "config.json";
+
+        private const string DIR_LOCALIZATION = "Localization";
+        private const string DEFAULT_LOCALIZATION = "en-US";
+
+        private const string DIR_PROFILES = "Profiles";
         private const string PROFILE_EXT = ".ckgprofile";
         private const int MAX_PROFILE_COUNT = 50;
 
         private static int _lastLoadedNumber = -1;
+
+        #region Public Config Functions
+
+        public static Config LoadConfig()
+        {
+            string path = GetConfigPath();
+            Config config = null;
+
+            JsonSerializerOptions options = new JsonSerializerOptions
+            {
+                IncludeFields = true,
+                WriteIndented = true
+            };
+
+            if (File.Exists(path) == false) //Create new
+            {
+                config = new Config();
+                config.LanguageCode = CultureInfo.CurrentUICulture.Name;
+
+                File.WriteAllText(path, JsonSerializer.Serialize(config, options));
+                return config;
+            }
+
+            string content = File.ReadAllText(path);
+            return JsonSerializer.Deserialize<Config>(content, options);
+        }
+
+        #endregion
+
+        #region Public Localization Functions
+
+        public static JsonElement LoadLocalization(string code)
+        {
+            string path = GetLocalizationPath(code);
+
+            if (File.Exists(path) == false)
+            {
+                //Load default language if system language is not supported
+                path = GetLocalizationPath(DEFAULT_LOCALIZATION);
+            }
+
+            string content = File.ReadAllText(path);
+            using JsonDocument document = JsonDocument.Parse(content);
+
+            return document.RootElement.Clone();
+        }
+
+        #endregion
+
+        #region Public Profile Functions
 
         public static UserProfile LoadDefaultProfile()
         {
@@ -28,7 +85,7 @@ namespace CKG
 
             string path = GetProfilePath(number);
 
-            if (File.Exists(path) == false)
+            if (File.Exists(path) == false) //Create new
             {
                 UserProfile.Current = new UserProfile();
                 UserProfile.Current.ProfileName = number == 1 ? "Default" : "New Profile";
@@ -67,8 +124,9 @@ namespace CKG
 
         public static List<SProfileInfo> GetProfileList()
         {
-            string dir = GetDirectoryPath();
+            string dir = GetDirectoryPath(DIR_PROFILES);
             List<SProfileInfo> result = new List<SProfileInfo>();
+
             JsonSerializerOptions options = new JsonSerializerOptions
             {
                 IncludeFields = true,
@@ -92,23 +150,40 @@ namespace CKG
             return result;
         }
 
-        private static string GetProfilePath(int number)
-        {
-            return Path.Combine(GetDirectoryPath(), GetProfileName(number));
-        }
+        #endregion
 
-        private static string GetDirectoryPath()
+        #region Private Functions
+
+        private static string GetDirectoryPath(string dirName)
         {
             string baseDir = AppDomain.CurrentDomain.BaseDirectory;
-            string profileDir = Path.Combine(baseDir, "Profiles");
+            string profileDir = Path.Combine(baseDir, dirName);
 
             Directory.CreateDirectory(profileDir);
             return profileDir;
+        }
+
+        private static string GetConfigPath()
+        {
+            string baseDir = AppDomain.CurrentDomain.BaseDirectory;
+            return Path.Combine(baseDir, CONFIG_FILE_NAME);
+        }
+
+        private static string GetLocalizationPath(string code)
+        {
+            return Path.Combine(GetDirectoryPath(DIR_LOCALIZATION), $"{code}.json");
+        }
+
+        private static string GetProfilePath(int number)
+        {
+            return Path.Combine(GetDirectoryPath(DIR_PROFILES), GetProfileName(number));
         }
 
         private static string GetProfileName(int number)
         {
             return $"profile{number}{PROFILE_EXT}";
         }
+
+        #endregion
     }
 }