+++ Chattokouken Ganbarunoda/Global/Config.cs
@@ -1,0 +1,8 @@
+﻿
+namespace CKG
+{
+    public class Config
+    {
+        public string LanguageCode;
+    }
+}