+++ Chattokouken Ganbarunoda/Global/EInputMethod.cs
@@ -1,0 +1,9 @@
+﻿
+namespace CKG
+{
+    public enum EInputMethod
+    {
+        DirectInput,
+        OverlayInput
+    }
+}